fix: prerender error 수정
