+++ src/components/pages/KakaoAuth/index.ts
@@ -1,0 +1,1 @@
+export { KakaoAuth } from "./KakaoAuth";
